// ui/src/app/models/gene-set.ts
export interface GeneSet {
    abbreviation: string;
    attribution: string;
    count: number;
    created: string;
    curation_id: number;
    description: string
    file_id: number;
    gene_id_type: number;
    id: number;
    name: string;
    publication_abstract: string;
    publication_authors: string;
    publication_id: number;
    publication_journal: string;
    publication_month: string;
    publication_pages: string;
    publication_pubmed_id: string;
    publication_title: string;
    publication_volume: string;
    publication_year: string;
    score_type: number;
    species_id: number;
    status: string;
    threshold: string;
    updated: string;
    user_id: number;
}